+++ crates/router/src/connector/braintree.rs
@@ -1,4 +1,3 @@
-pub mod braintree_graphql_transformers;
 pub mod transformers;
 use std::{fmt::Debug, str::FromStr};
 
@@ -7,7 +6,7 @@ use base64::Engine;
 use common_utils::{crypto, ext_traits::XmlExt, request::RequestContent};
 use diesel_models::enums;
 use error_stack::{report, Report, ResultExt};
-use masking::{ExposeInterface, PeekInterface};
+use masking::{ExposeInterface};
 use ring::hmac;
 use sha1::{Digest, Sha1};
 
@@ -43,12 +42,6 @@ pub struct Braintree;
 pub const BRAINTREE_VERSION: &str = "Braintree-Version";
 pub const BRAINTREE_VERSION_VALUE: &str = "2019-01-01";
 pub const BRAINTREE_API_VERSION: &str = "graphql_api";
-
-impl Braintree {
-    pub fn is_braintree_graphql_version(&self, connector_api_version: &Option<String>) -> bool {
-        *connector_api_version == Some(BRAINTREE_API_VERSION.to_string())
-    }
-}
 
 impl<Flow, Request, Response> ConnectorCommonExt<Flow, Request, Response> for Braintree
 where
@@ -105,11 +98,11 @@ impl ConnectorCommon for Braintree {
         res: types::Response,
         event_builder: Option<&mut ConnectorEvent>,
     ) -> CustomResult<ErrorResponse, errors::ConnectorError> {
-        let response: Result<braintree::ErrorResponse, Report<common_utils::errors::ParsingError>> =
+        let response: Result<braintree::ErrorResponses, Report<common_utils::errors::ParsingError>> =
             res.response.parse_struct("Braintree Error Response");
 
         match response {
-            Ok(braintree::ErrorResponse::BraintreeApiErrorResponse(response)) => {
+            Ok(braintree::ErrorResponses::BraintreeApiErrorResponse(response)) => {
                 event_builder.map(|i| i.set_error_response_body(&response));
                 router_env::logger::info!(connector_response=?response);
 
@@ -139,7 +132,7 @@ impl ConnectorCommon for Braintree {
                     connector_transaction_id: None,
                 })
             }
-            Ok(braintree::ErrorResponse::BraintreeErrorResponse(response)) => {
+            Ok(braintree::ErrorResponses::BraintreeErrorResponse(response)) => {
                 event_builder.map(|i| i.set_error_response_body(&response));
                 router_env::logger::info!(connector_response=?response);
 
@@ -196,109 +189,7 @@ impl ConnectorIntegration<api::AccessTokenAuth, types::AccessTokenRequestData, t
 impl ConnectorIntegration<api::Session, types::PaymentsSessionData, types::PaymentsResponseData>
     for Braintree
 {
-    fn get_headers(
-        &self,
-        req: &types::PaymentsSessionRouterData,
-        _connectors: &settings::Connectors,
-    ) -> CustomResult<Vec<(String, request::Maskable<String>)>, errors::ConnectorError> {
-        let mut headers = vec![
-            (
-                headers::CONTENT_TYPE.to_string(),
-                types::PaymentsSessionType::get_content_type(self)
-                    .to_string()
-                    .into(),
-            ),
-            (headers::X_API_VERSION.to_string(), "6".to_string().into()),
-            (
-                headers::ACCEPT.to_string(),
-                "application/json".to_string().into(),
-            ),
-        ];
-        let mut api_key = self.get_auth_header(&req.connector_auth_type)?;
-        headers.append(&mut api_key);
-        Ok(headers)
-    }
-
-    fn get_content_type(&self) -> &'static str {
-        "application/json"
-    }
-
-    fn get_url(
-        &self,
-        req: &types::PaymentsSessionRouterData,
-        connectors: &settings::Connectors,
-    ) -> CustomResult<String, errors::ConnectorError> {
-        let auth_type = braintree::BraintreeAuthType::try_from(&req.connector_auth_type)
-            .change_context(errors::ConnectorError::FailedToObtainAuthType)?;
-        Ok(format!(
-            "{}/merchants/{}/client_token",
-            self.base_url(connectors),
-            auth_type.merchant_id.peek(),
-        ))
-    }
-
-    fn build_request(
-        &self,
-        req: &types::PaymentsSessionRouterData,
-        connectors: &settings::Connectors,
-    ) -> CustomResult<Option<services::Request>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => Ok(None),
-            false => {
-                let request = Some(
-                    services::RequestBuilder::new()
-                        .method(services::Method::Post)
-                        .url(&types::PaymentsSessionType::get_url(self, req, connectors)?)
-                        .attach_default_headers()
-                        .headers(types::PaymentsSessionType::get_headers(
-                            self, req, connectors,
-                        )?)
-                        .set_body(types::PaymentsSessionType::get_request_body(
-                            self, req, connectors,
-                        )?)
-                        .build(),
-                );
-                Ok(request)
-            }
-        }
-    }
-
-    fn get_error_response(
-        &self,
-        res: types::Response,
-        event_builder: Option<&mut ConnectorEvent>,
-    ) -> CustomResult<ErrorResponse, errors::ConnectorError> {
-        self.build_error_response(res, event_builder)
-    }
-
-    fn get_request_body(
-        &self,
-        req: &types::PaymentsSessionRouterData,
-        _connectors: &settings::Connectors,
-    ) -> CustomResult<RequestContent, errors::ConnectorError> {
-        let connector_req = braintree::BraintreeSessionRequest::try_from(req)?;
-        Ok(RequestContent::Json(Box::new(connector_req)))
-    }
-
-    fn handle_response(
-        &self,
-        data: &types::PaymentsSessionRouterData,
-        event_builder: Option<&mut ConnectorEvent>,
-        res: types::Response,
-    ) -> CustomResult<types::PaymentsSessionRouterData, errors::ConnectorError> {
-        let response: braintree::BraintreeSessionTokenResponse = res
-            .response
-            .parse_struct("braintree SessionTokenResponse")
-            .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
-        event_builder.map(|i| i.set_response_body(&response));
-        router_env::logger::info!(connector_response=?response);
-        types::RouterData::try_from(types::ResponseRouterData {
-            response,
-            data: data.clone(),
-            http_code: res.status_code,
-        })
-    }
+   
 }
 
 impl api::PaymentToken for Braintree {}
@@ -340,7 +231,7 @@ impl
         req: &types::TokenizationRouterData,
         _connectors: &settings::Connectors,
     ) -> CustomResult<RequestContent, errors::ConnectorError> {
-        let connector_req = braintree_graphql_transformers::BraintreeTokenRequest::try_from(req)?;
+        let connector_req = transformers::BraintreeTokenRequest::try_from(req)?;
 
         Ok(RequestContent::Json(Box::new(connector_req)))
     }
@@ -350,9 +241,7 @@ impl
         req: &types::TokenizationRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<Option<services::Request>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => Ok(Some(
+        Ok(Some(
                 services::RequestBuilder::new()
                     .method(services::Method::Post)
                     .url(&types::TokenizationType::get_url(self, req, connectors)?)
@@ -362,11 +251,8 @@ impl
                         self, req, connectors,
                     )?)
                     .build(),
-            )),
-            false => Ok(None),
+            ))
         }
-    }
-
     fn handle_response(
         &self,
         data: &types::TokenizationRouterData,
@@ -376,7 +262,7 @@ impl
     where
         types::PaymentsResponseData: Clone,
     {
-        let response: braintree_graphql_transformers::BraintreeTokenResponse = res
+        let response: transformers::BraintreeTokenResponse = res
             .response
             .parse_struct("BraintreeTokenResponse")
             .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
@@ -433,11 +319,7 @@ impl ConnectorIntegration<api::Capture, types::PaymentsCaptureData, types::Payme
         req: &types::PaymentsCaptureRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<Vec<(String, request::Maskable<String>)>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version.clone();
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => self.build_headers(req, connectors),
-            false => Ok(vec![]),
-        }
+        self.build_headers(req, connectors)
     }
 
     fn get_content_type(&self) -> &'static str {
@@ -446,62 +328,43 @@ impl ConnectorIntegration<api::Capture, types::PaymentsCaptureData, types::Payme
 
     fn get_url(
         &self,
-        req: &types::PaymentsCaptureRouterData,
+        _req: &types::PaymentsCaptureRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<String, errors::ConnectorError> {
-        let connector_api_version = req.connector_api_version.clone();
-        match self.is_braintree_graphql_version(&connector_api_version) {
-            true => {
-                let base_url = connectors
-                    .braintree
-                    .secondary_base_url
-                    .as_ref()
-                    .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
-                Ok(base_url.to_string())
-            }
-            false => {
-                Err(errors::ConnectorError::NotImplemented("get_url method".to_string()).into())
-            }
+        let base_url = connectors
+            .braintree
+            .secondary_base_url
+            .as_ref()
+            .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
+        Ok(base_url.to_string())
         }
-    }
 
     fn get_request_body(
         &self,
         req: &types::PaymentsCaptureRouterData,
         _connectors: &settings::Connectors,
     ) -> CustomResult<RequestContent, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version.clone();
         let connector_router_data =
-            braintree_graphql_transformers::BraintreeRouterData::try_from((
+            transformers::BraintreeRouterData::try_from((
                 &self.get_currency_unit(),
                 req.request.currency,
                 req.request.amount_to_capture,
                 req,
             ))?;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let connector_req =
-                    braintree_graphql_transformers::BraintreeCaptureRequest::try_from(
-                        &connector_router_data,
-                    )?;
+            let connector_req =
+                transformers::BraintreeCaptureRequest::try_from(
+                    &connector_router_data,
+                )?;
 
-                Ok(RequestContent::Json(Box::new(connector_req)))
-            }
-            false => Err(errors::ConnectorError::NotImplemented(
-                "get_request_body method".to_string(),
-            )
-            .into()),
+            Ok(RequestContent::Json(Box::new(connector_req)))
         }
-    }
 
     fn build_request(
         &self,
         req: &types::PaymentsCaptureRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<Option<services::Request>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => Ok(Some(
+            Ok(Some(
                 services::RequestBuilder::new()
                     .method(services::Method::Post)
                     .url(&types::PaymentsCaptureType::get_url(self, req, connectors)?)
@@ -513,13 +376,8 @@ impl ConnectorIntegration<api::Capture, types::PaymentsCaptureData, types::Payme
                         self, req, connectors,
                     )?)
                     .build(),
-            )),
-            false => Err(errors::ConnectorError::NotImplemented(
-                "Capture flow not Implemented".to_string(),
-            )
-            .into()),
+            ))
         }
-    }
 
     fn handle_response(
         &self,
@@ -527,7 +385,7 @@ impl ConnectorIntegration<api::Capture, types::PaymentsCaptureData, types::Payme
         event_builder: Option<&mut ConnectorEvent>,
         res: types::Response,
     ) -> CustomResult<types::PaymentsCaptureRouterData, errors::ConnectorError> {
-        let response: braintree_graphql_transformers::BraintreeCaptureResponse = res
+        let response: transformers::BraintreeCaptureResponse = res
             .response
             .parse_struct("Braintree PaymentsCaptureResponse")
             .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
@@ -557,29 +415,8 @@ impl ConnectorIntegration<api::PSync, types::PaymentsSyncData, types::PaymentsRe
         req: &types::PaymentsSyncRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<Vec<(String, request::Maskable<String>)>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => self.build_headers(req, connectors),
-            false => {
-                let mut headers = vec![
-                    (
-                        headers::CONTENT_TYPE.to_string(),
-                        types::PaymentsSyncType::get_content_type(self)
-                            .to_string()
-                            .into(),
-                    ),
-                    (headers::X_API_VERSION.to_string(), "6".to_string().into()),
-                    (
-                        headers::ACCEPT.to_string(),
-                        "application/json".to_string().into(),
-                    ),
-                ];
-                let mut api_key = self.get_auth_header(&req.connector_auth_type)?;
-                headers.append(&mut api_key);
-                Ok(headers)
-            }
+            self.build_headers(req, connectors)
         }
-    }
 
     fn get_content_type(&self) -> &'static str {
         self.common_get_content_type()
@@ -587,61 +424,34 @@ impl ConnectorIntegration<api::PSync, types::PaymentsSyncData, types::PaymentsRe
 
     fn get_url(
         &self,
-        req: &types::PaymentsSyncRouterData,
+        _req: &types::PaymentsSyncRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<String, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let base_url = connectors
-                    .braintree
-                    .secondary_base_url
-                    .as_ref()
-                    .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
-                Ok(base_url.to_string())
-            }
-            false => {
-                let auth_type = braintree::BraintreeAuthType::try_from(&req.connector_auth_type)
-                    .change_context(errors::ConnectorError::FailedToObtainAuthType)?;
-                let connector_payment_id = req
-                    .request
-                    .connector_transaction_id
-                    .get_connector_transaction_id()
-                    .change_context(errors::ConnectorError::MissingConnectorTransactionID)?;
-                Ok(format!(
-                    "{}/merchants/{}/transactions/{connector_payment_id}",
-                    self.base_url(connectors),
-                    auth_type.merchant_id.peek()
-                ))
-            }
+        let base_url = connectors
+            .braintree
+            .secondary_base_url
+            .as_ref()
+            .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
+        Ok(base_url.to_string())
         }
-    }
 
     fn get_request_body(
         &self,
         req: &types::PaymentsSyncRouterData,
         _connectors: &settings::Connectors,
     ) -> CustomResult<RequestContent, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let connector_req =
-                    braintree_graphql_transformers::BraintreePSyncRequest::try_from(req)?;
+        let connector_req =
+            transformers::BraintreePSyncRequest::try_from(req)?;
 
-                Ok(RequestContent::Json(Box::new(connector_req)))
+        Ok(RequestContent::Json(Box::new(connector_req)))
             }
-            false => Err(report!(errors::ConnectorError::RequestEncodingFailed)),
-        }
-    }
 
     fn build_request(
         &self,
         req: &types::PaymentsSyncRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<Option<services::Request>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => Ok(Some(
+            Ok(Some(
                 services::RequestBuilder::new()
                     .method(services::Method::Post)
                     .url(&types::PaymentsSyncType::get_url(self, req, connectors)?)
@@ -651,20 +461,8 @@ impl ConnectorIntegration<api::PSync, types::PaymentsSyncData, types::PaymentsRe
                         self, req, connectors,
                     )?)
                     .build(),
-            )),
-            false => Ok(Some(
-                services::RequestBuilder::new()
-                    .method(services::Method::Get)
-                    .url(&types::PaymentsSyncType::get_url(self, req, connectors)?)
-                    .attach_default_headers()
-                    .headers(types::PaymentsSyncType::get_headers(self, req, connectors)?)
-                    .set_body(types::PaymentsSyncType::get_request_body(
-                        self, req, connectors,
-                    )?)
-                    .build(),
-            )),
+            ))
         }
-    }
 
     fn handle_response(
         &self,
@@ -672,36 +470,19 @@ impl ConnectorIntegration<api::PSync, types::PaymentsSyncData, types::PaymentsRe
         event_builder: Option<&mut ConnectorEvent>,
         res: types::Response,
     ) -> CustomResult<types::PaymentsSyncRouterData, errors::ConnectorError> {
-        let connector_api_version = &data.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let response: braintree_graphql_transformers::BraintreePSyncResponse = res
-                    .response
-                    .parse_struct("Braintree PaymentSyncResponse")
-                    .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
-                event_builder.map(|i| i.set_response_body(&response));
-                router_env::logger::info!(connector_response=?response);
-                types::RouterData::try_from(types::ResponseRouterData {
-                    response,
-                    data: data.clone(),
-                    http_code: res.status_code,
-                })
+            
+            let response: transformers::BraintreePSyncResponse = res
+                .response
+                .parse_struct("Braintree PaymentSyncResponse")
+                .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
+            event_builder.map(|i| i.set_response_body(&response));
+            router_env::logger::info!(connector_response=?response);
+            types::RouterData::try_from(types::ResponseRouterData {
+                response,
+                data: data.clone(),
+                http_code: res.status_code,
+            })
             }
-            false => {
-                let response: braintree::BraintreePaymentsResponse = res
-                    .response
-                    .parse_struct("Braintree PaymentsResponse")
-                    .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
-                event_builder.map(|i| i.set_response_body(&response));
-                router_env::logger::info!(connector_response=?response);
-                types::RouterData::try_from(types::ResponseRouterData {
-                    response,
-                    data: data.clone(),
-                    http_code: res.status_code,
-                })
-            }
-        }
-    }
 
     fn get_error_response(
         &self,
@@ -720,56 +501,20 @@ impl ConnectorIntegration<api::Authorize, types::PaymentsAuthorizeData, types::P
         req: &types::PaymentsAuthorizeRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<Vec<(String, request::Maskable<String>)>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => self.build_headers(req, connectors),
-            false => {
-                let mut headers = vec![
-                    (
-                        headers::CONTENT_TYPE.to_string(),
-                        types::PaymentsAuthorizeType::get_content_type(self)
-                            .to_string()
-                            .into(),
-                    ),
-                    (headers::X_API_VERSION.to_string(), "6".to_string().into()),
-                    (
-                        headers::ACCEPT.to_string(),
-                        "application/json".to_string().into(),
-                    ),
-                ];
-                let mut api_key = self.get_auth_header(&req.connector_auth_type)?;
-                headers.append(&mut api_key);
-                Ok(headers)
-            }
-        }
+        self.build_headers(req, connectors)
     }
 
     fn get_url(
         &self,
-        req: &types::PaymentsAuthorizeRouterData,
+        _req: &types::PaymentsAuthorizeRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<String, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let base_url = connectors
-                    .braintree
-                    .secondary_base_url
-                    .as_ref()
-                    .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
-                Ok(base_url.to_string())
-            }
-            false => {
-                let auth_type = braintree::BraintreeAuthType::try_from(&req.connector_auth_type)
-                    .change_context(errors::ConnectorError::FailedToObtainAuthType)?;
-
-                Ok(format!(
-                    "{}merchants/{}/transactions",
-                    self.base_url(connectors),
-                    auth_type.merchant_id.peek()
-                ))
-            }
-        }
+        let base_url = connectors
+            .braintree
+            .secondary_base_url
+            .as_ref()
+            .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
+        Ok(base_url.to_string())
     }
 
     fn build_request(
@@ -799,27 +544,18 @@ impl ConnectorIntegration<api::Authorize, types::PaymentsAuthorizeData, types::P
         req: &types::PaymentsAuthorizeRouterData,
         _connectors: &settings::Connectors,
     ) -> CustomResult<RequestContent, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
         let connector_router_data =
-            braintree_graphql_transformers::BraintreeRouterData::try_from((
+            transformers::BraintreeRouterData::try_from((
                 &self.get_currency_unit(),
                 req.request.currency,
                 req.request.amount,
                 req,
             ))?;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let connector_req =
-                    braintree_graphql_transformers::BraintreePaymentsRequest::try_from(
-                        &connector_router_data,
-                    )?;
-                Ok(RequestContent::Json(Box::new(connector_req)))
-            }
-            false => {
-                let connector_req = braintree::BraintreePaymentsRequest::try_from(req)?;
-                Ok(RequestContent::Json(Box::new(connector_req)))
-            }
-        }
+            let connector_req: transformers::BraintreePaymentsRequest =
+                transformers::BraintreePaymentsRequest::try_from(
+                    &connector_router_data,
+                )?;
+            Ok(RequestContent::Json(Box::new(connector_req)))
     }
 
     fn handle_response(
@@ -828,11 +564,9 @@ impl ConnectorIntegration<api::Authorize, types::PaymentsAuthorizeData, types::P
         event_builder: Option<&mut ConnectorEvent>,
         res: types::Response,
     ) -> CustomResult<types::PaymentsAuthorizeRouterData, errors::ConnectorError> {
-        let connector_api_version = &data.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => match data.request.is_auto_capture()? {
+        match data.request.is_auto_capture()? {
                 true => {
-                    let response: braintree_graphql_transformers::BraintreePaymentsResponse = res
+                    let response: transformers::BraintreePaymentsResponse = res
                         .response
                         .parse_struct("Braintree PaymentsResponse")
                         .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
@@ -845,7 +579,7 @@ impl ConnectorIntegration<api::Authorize, types::PaymentsAuthorizeData, types::P
                     })
                 }
                 false => {
-                    let response: braintree_graphql_transformers::BraintreeAuthResponse = res
+                    let response: transformers::BraintreeAuthResponse = res
                         .response
                         .parse_struct("Braintree AuthResponse")
                         .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
@@ -857,22 +591,8 @@ impl ConnectorIntegration<api::Authorize, types::PaymentsAuthorizeData, types::P
                         http_code: res.status_code,
                     })
                 }
-            },
-            false => {
-                let response: braintree::BraintreePaymentsResponse = res
-                    .response
-                    .parse_struct("Braintree PaymentsResponse")
-                    .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
-                event_builder.map(|i| i.set_response_body(&response));
-                router_env::logger::info!(connector_response=?response);
-                types::RouterData::try_from(types::ResponseRouterData {
-                    response,
-                    data: data.clone(),
-                    http_code: res.status_code,
-                })
             }
         }
-    }
 
     fn get_error_response(
         &self,
@@ -891,28 +611,7 @@ impl ConnectorIntegration<api::Void, types::PaymentsCancelData, types::PaymentsR
         req: &types::PaymentsCancelRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<Vec<(String, request::Maskable<String>)>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => self.build_headers(req, connectors),
-            false => {
-                let mut headers = vec![
-                    (
-                        headers::CONTENT_TYPE.to_string(),
-                        types::PaymentsVoidType::get_content_type(self)
-                            .to_string()
-                            .into(),
-                    ),
-                    (headers::X_API_VERSION.to_string(), "6".to_string().into()),
-                    (
-                        headers::ACCEPT.to_string(),
-                        "application/json".to_string().into(),
-                    ),
-                ];
-                let mut api_key = self.get_auth_header(&req.connector_auth_type)?;
-                headers.append(&mut api_key);
-                Ok(headers)
-            }
-        }
+        self.build_headers(req, connectors)
     }
 
     fn get_content_type(&self) -> &'static str {
@@ -921,31 +620,16 @@ impl ConnectorIntegration<api::Void, types::PaymentsCancelData, types::PaymentsR
 
     fn get_url(
         &self,
-        req: &types::PaymentsCancelRouterData,
+        _req: &types::PaymentsCancelRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<String, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let base_url = connectors
+        let base_url = connectors
                     .braintree
                     .secondary_base_url
                     .as_ref()
                     .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
                 Ok(base_url.to_string())
             }
-            false => {
-                let auth_type = braintree::BraintreeAuthType::try_from(&req.connector_auth_type)
-                    .change_context(errors::ConnectorError::FailedToObtainAuthType)?;
-                Ok(format!(
-                    "{}merchants/{}/transactions/{}/void",
-                    self.base_url(connectors),
-                    auth_type.merchant_id.peek(),
-                    req.request.connector_transaction_id
-                ))
-            }
-        }
-    }
 
     fn build_request(
         &self,
@@ -970,16 +654,10 @@ impl ConnectorIntegration<api::Void, types::PaymentsCancelData, types::PaymentsR
         req: &types::PaymentsCancelRouterData,
         _connectors: &settings::Connectors,
     ) -> CustomResult<RequestContent, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let connector_req =
-                    braintree_graphql_transformers::BraintreeCancelRequest::try_from(req)?;
-                Ok(RequestContent::Json(Box::new(connector_req)))
+        let connector_req =
+            transformers::BraintreeCancelRequest::try_from(req)?;
+        Ok(RequestContent::Json(Box::new(connector_req)))
             }
-            false => Err(report!(errors::ConnectorError::RequestEncodingFailed)),
-        }
-    }
 
     fn handle_response(
         &self,
@@ -987,10 +665,7 @@ impl ConnectorIntegration<api::Void, types::PaymentsCancelData, types::PaymentsR
         event_builder: Option<&mut ConnectorEvent>,
         res: types::Response,
     ) -> CustomResult<types::PaymentsCancelRouterData, errors::ConnectorError> {
-        let connector_api_version = &data.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let response: braintree_graphql_transformers::BraintreeCancelResponse = res
+        let response: transformers::BraintreeCancelResponse = res
                     .response
                     .parse_struct("Braintree VoidResponse")
                     .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
@@ -1002,22 +677,6 @@ impl ConnectorIntegration<api::Void, types::PaymentsCancelData, types::PaymentsR
                     http_code: res.status_code,
                 })
             }
-            false => {
-                let response: braintree::BraintreePaymentsResponse = res
-                    .response
-                    .parse_struct("Braintree PaymentsVoidResponse")
-                    .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
-                event_builder.map(|i| i.set_response_body(&response));
-                router_env::logger::info!(connector_response=?response);
-                types::RouterData::try_from(types::ResponseRouterData {
-                    response,
-                    data: data.clone(),
-                    http_code: res.status_code,
-                })
-            }
-        }
-    }
-
     fn get_error_response(
         &self,
         res: types::Response,
@@ -1039,28 +698,7 @@ impl ConnectorIntegration<api::Execute, types::RefundsData, types::RefundsRespon
         req: &types::RefundsRouterData<api::Execute>,
         connectors: &settings::Connectors,
     ) -> CustomResult<Vec<(String, request::Maskable<String>)>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => self.build_headers(req, connectors),
-            false => {
-                let mut headers = vec![
-                    (
-                        headers::CONTENT_TYPE.to_string(),
-                        types::RefundExecuteType::get_content_type(self)
-                            .to_string()
-                            .into(),
-                    ),
-                    (headers::X_API_VERSION.to_string(), "6".to_string().into()),
-                    (
-                        headers::ACCEPT.to_string(),
-                        "application/json".to_string().into(),
-                    ),
-                ];
-                let mut api_key = self.get_auth_header(&req.connector_auth_type)?;
-                headers.append(&mut api_key);
-                Ok(headers)
-            }
-        }
+        self.build_headers(req, connectors)
     }
 
     fn get_content_type(&self) -> &'static str {
@@ -1069,61 +707,35 @@ impl ConnectorIntegration<api::Execute, types::RefundsData, types::RefundsRespon
 
     fn get_url(
         &self,
-        req: &types::RefundsRouterData<api::Execute>,
+        _req: &types::RefundsRouterData<api::Execute>,
         connectors: &settings::Connectors,
     ) -> CustomResult<String, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let base_url = connectors
-                    .braintree
-                    .secondary_base_url
-                    .as_ref()
-                    .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
-                Ok(base_url.to_string())
+            let base_url = connectors
+                .braintree
+                .secondary_base_url
+                .as_ref()
+                .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
+            Ok(base_url.to_string())
             }
-            false => {
-                let auth_type = braintree::BraintreeAuthType::try_from(&req.connector_auth_type)
-                    .change_context(errors::ConnectorError::FailedToObtainAuthType)?;
-                let connector_payment_id = req.request.connector_transaction_id.clone();
-                Ok(format!(
-                    "{}merchants/{}/transactions/{}",
-                    self.base_url(connectors),
-                    auth_type.merchant_id.peek(),
-                    connector_payment_id
-                ))
-            }
-        }
-    }
 
     fn get_request_body(
         &self,
         req: &types::RefundsRouterData<api::Execute>,
         _connectors: &settings::Connectors,
     ) -> CustomResult<RequestContent, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
         let connector_router_data =
-            braintree_graphql_transformers::BraintreeRouterData::try_from((
+            transformers::BraintreeRouterData::try_from((
                 &self.get_currency_unit(),
                 req.request.currency,
                 req.request.refund_amount,
                 req,
             ))?;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let connector_req =
-                    braintree_graphql_transformers::BraintreeRefundRequest::try_from(
+        let connector_req =
+                    transformers::BraintreeRefundRequest::try_from(
                         connector_router_data,
                     )?;
-                Ok(RequestContent::Json(Box::new(connector_req)))
-            }
-            false => {
-                let connector_req = braintree::BraintreeRefundRequest::try_from(req)?;
-                Ok(RequestContent::Json(Box::new(connector_req)))
-            }
+        Ok(RequestContent::Json(Box::new(connector_req)))
         }
-    }
-
     fn build_request(
         &self,
         req: &types::RefundsRouterData<api::Execute>,
@@ -1149,36 +761,18 @@ impl ConnectorIntegration<api::Execute, types::RefundsData, types::RefundsRespon
         event_builder: Option<&mut ConnectorEvent>,
         res: types::Response,
     ) -> CustomResult<types::RefundsRouterData<api::Execute>, errors::ConnectorError> {
-        let connector_api_version = &data.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let response: braintree_graphql_transformers::BraintreeRefundResponse = res
-                    .response
-                    .parse_struct("Braintree RefundResponse")
-                    .change_context(errors::ConnectorError::RequestEncodingFailed)?;
-                event_builder.map(|i| i.set_response_body(&response));
-                router_env::logger::info!(connector_response=?response);
-                types::RouterData::try_from(types::ResponseRouterData {
-                    response,
-                    data: data.clone(),
-                    http_code: res.status_code,
-                })
+            let response: transformers::BraintreeRefundResponse = res
+                .response
+                .parse_struct("Braintree RefundResponse")
+                .change_context(errors::ConnectorError::RequestEncodingFailed)?;
+            event_builder.map(|i| i.set_response_body(&response));
+            router_env::logger::info!(connector_response=?response);
+            types::RouterData::try_from(types::ResponseRouterData {
+                response,
+                data: data.clone(),
+                http_code: res.status_code,
+            })
             }
-            false => {
-                let response: braintree::RefundResponse = res
-                    .response
-                    .parse_struct("Braintree RefundResponse")
-                    .change_context(errors::ConnectorError::RequestEncodingFailed)?;
-                event_builder.map(|i| i.set_response_body(&response));
-                router_env::logger::info!(connector_response=?response);
-                types::RouterData::try_from(types::ResponseRouterData {
-                    response,
-                    data: data.clone(),
-                    http_code: res.status_code,
-                })
-            }
-        }
-    }
 
     fn get_error_response(
         &self,
@@ -1197,12 +791,8 @@ impl ConnectorIntegration<api::RSync, types::RefundsData, types::RefundsResponse
         req: &types::RefundSyncRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<Vec<(String, request::Maskable<String>)>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => self.build_headers(req, connectors),
-            false => Ok(vec![]),
+            self.build_headers(req, connectors)
         }
-    }
 
     fn get_content_type(&self) -> &'static str {
         self.common_get_content_type()
@@ -1210,49 +800,33 @@ impl ConnectorIntegration<api::RSync, types::RefundsData, types::RefundsResponse
 
     fn get_url(
         &self,
-        req: &types::RefundSyncRouterData,
+        _req: &types::RefundSyncRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<String, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let base_url = connectors
+        let base_url = connectors
                     .braintree
                     .secondary_base_url
                     .as_ref()
                     .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
                 Ok(base_url.to_string())
-            }
-            false => {
-                Err(errors::ConnectorError::NotImplemented("get_url method".to_string()).into())
-            }
         }
-    }
 
     fn get_request_body(
         &self,
         req: &types::RefundSyncRouterData,
         _connectors: &settings::Connectors,
     ) -> CustomResult<RequestContent, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let connector_req =
-                    braintree_graphql_transformers::BraintreeRSyncRequest::try_from(req)?;
+            let connector_req =
+                    transformers::BraintreeRSyncRequest::try_from(req)?;
                 Ok(RequestContent::Json(Box::new(connector_req)))
             }
-            false => Err(report!(errors::ConnectorError::RequestEncodingFailed)),
-        }
-    }
 
     fn build_request(
         &self,
         req: &types::RefundSyncRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<Option<services::Request>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => Ok(Some(
+        Ok(Some(
                 services::RequestBuilder::new()
                     .method(services::Method::Post)
                     .url(&types::RefundSyncType::get_url(self, req, connectors)?)
@@ -1262,10 +836,8 @@ impl ConnectorIntegration<api::RSync, types::RefundsData, types::RefundsResponse
                         self, req, connectors,
                     )?)
                     .build(),
-            )),
-            false => Ok(None),
+            ))
         }
-    }
 
     fn handle_response(
         &self,
@@ -1276,36 +848,18 @@ impl ConnectorIntegration<api::RSync, types::RefundsData, types::RefundsResponse
         types::RouterData<api::RSync, types::RefundsData, types::RefundsResponseData>,
         errors::ConnectorError,
     > {
-        let connector_api_version = &data.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let response: braintree_graphql_transformers::BraintreeRSyncResponse = res
-                    .response
-                    .parse_struct("Braintree RefundResponse")
-                    .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
-                event_builder.map(|i| i.set_response_body(&response));
-                router_env::logger::info!(connector_response=?response);
-                types::RouterData::try_from(types::ResponseRouterData {
-                    response,
-                    data: data.clone(),
-                    http_code: res.status_code,
-                })
-            }
-            false => {
-                let response: braintree::RefundResponse = res
-                    .response
-                    .parse_struct("Braintree RefundResponse")
-                    .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
-                event_builder.map(|i| i.set_response_body(&response));
-                router_env::logger::info!(connector_response=?response);
-                types::RouterData::try_from(types::ResponseRouterData {
-                    response,
-                    data: data.clone(),
-                    http_code: res.status_code,
-                })
-            }
+        let response: transformers::BraintreeRSyncResponse = res
+            .response
+            .parse_struct("Braintree RefundResponse")
+            .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
+        event_builder.map(|i| i.set_response_body(&response));
+        router_env::logger::info!(connector_response=?response);
+        types::RouterData::try_from(types::ResponseRouterData {
+            response,
+            data: data.clone(),
+            http_code: res.status_code,
+        })
         }
-    }
     fn get_error_response(
         &self,
         res: types::Response,
@@ -1475,7 +1029,7 @@ impl api::IncomingWebhook for Braintree {
                         currency,
                     )?,
                     currency: dispute_data.currency_iso_code,
-                    dispute_stage: braintree_graphql_transformers::get_dispute_stage(
+                    dispute_stage: transformers::get_dispute_stage(
                         dispute_data.kind.as_str(),
                     )?,
                     connector_dispute_id: dispute_data.id,
@@ -1506,8 +1060,8 @@ fn get_matching_webhook_signature(
 
 fn get_webhook_object_from_body(
     body: &[u8],
-) -> CustomResult<braintree_graphql_transformers::BraintreeWebhookResponse, errors::ParsingError> {
-    serde_urlencoded::from_bytes::<braintree_graphql_transformers::BraintreeWebhookResponse>(body)
+) -> CustomResult<transformers::BraintreeWebhookResponse, errors::ParsingError> {
+    serde_urlencoded::from_bytes::<transformers::BraintreeWebhookResponse>(body)
         .change_context(errors::ParsingError::StructParseFailure(
             "BraintreeWebhookResponse",
         ))
@@ -1515,7 +1069,7 @@ fn get_webhook_object_from_body(
 
 fn decode_webhook_payload(
     payload: &[u8],
-) -> CustomResult<braintree_graphql_transformers::Notification, errors::ConnectorError> {
+) -> CustomResult<transformers::Notification, errors::ConnectorError> {
     let decoded_response = consts::BASE64_ENGINE
         .decode(payload)
         .change_context(errors::ConnectorError::WebhookBodyDecodingFailed)?;
@@ -1524,7 +1078,7 @@ fn decode_webhook_payload(
         .change_context(errors::ConnectorError::WebhookBodyDecodingFailed)?;
 
     xml_response
-        .parse_xml::<braintree_graphql_transformers::Notification>()
+        .parse_xml::<transformers::Notification>()
         .change_context(errors::ConnectorError::WebhookBodyDecodingFailed)
 }
 
@@ -1538,7 +1092,7 @@ impl services::ConnectorRedirectResponse for Braintree {
         match action {
             services::PaymentAction::PSync => match json_payload {
                 Some(payload) => {
-                    let redirection_response:braintree_graphql_transformers::BraintreeRedirectionResponse = serde_json::from_value(payload)
+                    let redirection_response:transformers::BraintreeRedirectionResponse = serde_json::from_value(payload)
 
                             .change_context(
                                 errors::ConnectorError::MissingConnectorRedirectionPayload {
@@ -1547,7 +1101,7 @@ impl services::ConnectorRedirectResponse for Braintree {
                             )?;
                     let braintree_payload =
                         serde_json::from_str::<
-                            braintree_graphql_transformers::BraintreeThreeDsErrorResponse,
+                            transformers::BraintreeThreeDsErrorResponse,
                         >(&redirection_response.authentication_response);
                     let (error_code, error_message) = match braintree_payload {
                         Ok(braintree_response_payload) => (
@@ -1587,72 +1141,49 @@ impl
         req: &types::PaymentsCompleteAuthorizeRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<Vec<(String, request::Maskable<String>)>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => self.build_headers(req, connectors),
-            false => Err(errors::ConnectorError::NotImplemented(
-                "get_headers method".to_string(),
-            ))?,
+        self.build_headers(req, connectors)
         }
-    }
     fn get_content_type(&self) -> &'static str {
         self.common_get_content_type()
     }
     fn get_url(
         &self,
-        req: &types::PaymentsCompleteAuthorizeRouterData,
+        _req: &types::PaymentsCompleteAuthorizeRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<String, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let base_url = connectors
-                    .braintree
-                    .secondary_base_url
-                    .as_ref()
-                    .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
-                Ok(base_url.to_string())
-            }
-            false => Err(errors::ConnectorError::NotImplemented(
-                "get_url method".to_string(),
-            ))?,
-        }
-    }
+        let base_url = connectors
+            .braintree
+            .secondary_base_url
+            .as_ref()
+            .ok_or(errors::ConnectorError::FailedToObtainIntegrationUrl)?;
+        Ok(base_url.to_string())
+}
     fn get_request_body(
         &self,
         req: &types::PaymentsCompleteAuthorizeRouterData,
         _connectors: &settings::Connectors,
     ) -> CustomResult<RequestContent, errors::ConnectorError> {
         let connector_router_data =
-            braintree_graphql_transformers::BraintreeRouterData::try_from((
+            transformers::BraintreeRouterData::try_from((
                 &self.get_currency_unit(),
                 req.request.currency,
                 req.request.amount,
                 req,
             ))?;
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => {
-                let connector_req =
-                    braintree_graphql_transformers::BraintreePaymentsRequest::try_from(
-                        &connector_router_data,
-                    )?;
-                Ok(RequestContent::Json(Box::new(connector_req)))
-            }
-            false => Err(errors::ConnectorError::NotImplemented(
-                "get_request_body method".to_string(),
-            ))?,
+            
+        let connector_req =
+            transformers::BraintreePaymentsRequest::try_from(
+                &connector_router_data,
+            )?;
+        Ok(RequestContent::Json(Box::new(connector_req)))
         }
-    }
 
     fn build_request(
         &self,
         req: &types::PaymentsCompleteAuthorizeRouterData,
         connectors: &settings::Connectors,
     ) -> CustomResult<Option<services::Request>, errors::ConnectorError> {
-        let connector_api_version = &req.connector_api_version;
-        match self.is_braintree_graphql_version(connector_api_version) {
-            true => Ok(Some(
+            Ok(Some(
                 services::RequestBuilder::new()
                     .method(services::Method::Post)
                     .url(&types::PaymentsCompleteAuthorizeType::get_url(
@@ -1666,11 +1197,7 @@ impl
                         self, req, connectors,
                     )?)
                     .build(),
-            )),
-            false => Err(errors::ConnectorError::NotImplemented(
-                "payment method".to_string(),
-            ))?,
-        }
+            ))
     }
     fn handle_response(
         &self,
@@ -1681,7 +1208,7 @@ impl
         match connector_utils::PaymentsCompleteAuthorizeRequestData::is_auto_capture(&data.request)?
         {
             true => {
-                let response: braintree_graphql_transformers::BraintreeCompleteChargeResponse = res
+                let response: transformers::BraintreeCompleteChargeResponse = res
                     .response
                     .parse_struct("Braintree PaymentsResponse")
                     .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
@@ -1694,7 +1221,7 @@ impl
                 })
             }
             false => {
-                let response: braintree_graphql_transformers::BraintreeCompleteAuthResponse = res
+                let response: transformers::BraintreeCompleteAuthResponse = res
                     .response
                     .parse_struct("Braintree AuthResponse")
                     .change_context(errors::ConnectorError::ResponseDeserializationFailed)?;
